# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApplyInvoicePromoRequest": {
        "additionalProperties": false,
        "description": "Apply a case-insensitive code, or remove it with explicit null.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "required": [
          "code"
        ],
        "title": "ApplyInvoicePromoRequest",
        "type": "object"
      },
...
      "CreateInvoiceRequest": {
        "additionalProperties": false,
        "description": "Author and issue an invoice for an order.",
        "properties": {
          "courier_fee_amount": {
            "default": "0.00",
            "description": "Courier's craft/labour fee.",
            "maxLength": 20,
            "minLength": 1,
            "title": "Courier Fee Amount",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineRequest"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "promo_code": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promo Code"
          }
        },
        "required": [
          "items"
        ],
        "title": "CreateInvoiceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceLineRequest": {
        "additionalProperties": false,
        "description": "One courier-authored invoice line with its final unit price.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "quantity": {
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "unit_price_amount": {
            "description": "Final unit price, e.g. \"400.00\".",
            "maxLength": 20,
            "minLength": 1,
            "title": "Unit Price Amount",
            "type": "string"
          }
        },
        "required": [
          "title",
          "unit_price_amount",
          "quantity"
        ],
        "title": "InvoiceLineRequest",
        "type": "object"
      },
      "InvoicePage": {
        "description": "Newest-created invoices first, with an owned UUID cursor.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "InvoicePage",
        "type": "object"
      },
      "InvoiceResponse": {
        "description": "A full invoice view for its participants (every amount server-computed).",
        "properties": {
          "courier_fee_amount": {
            "title": "Courier Fee Amount",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "discount_amount": {
            "title": "Discount Amount",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "items_net_amount": {
            "title": "Items Net Amount",
            "type": "string"
          },
          "net_after_discount_amount": {
            "title": "Net After Discount Amount",
            "type": "string"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "promo_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promo Code"
          },
          "service_fee_amount": {
            "title": "Service Fee Amount",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_amount": {
            "title": "Total Amount",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_id",
          "status",
          "currency",
          "items_net_amount",
          "courier_fee_amount",
          "service_fee_amount",
          "discount_amount",
          "net_after_discount_amount",
          "total_amount",
          "promo_code",
          "issued_at",
          "expires_at",
          "items"
        ],
        "title": "InvoiceResponse",
        "type": "object"
      },
      "InvoiceStatus": {
        "description": "Invoice lifecycle status.",
        "enum": [
          "DRAFT",
          "ISSUED",
          "PAID",
          "CANCELLED",
          "EXPIRED",
          "REFUNDED"
        ],
        "title": "InvoiceStatus",
        "type": "string"
      },
      "InvoiceSummary": {
        "description": "One visible invoice revision and its current-revision marker.",
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "issued_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "total_amount": {
            "title": "Total Amount",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_id",
          "status",
          "currency",
          "total_amount",
          "issued_at",
          "expires_at",
          "is_current"
        ],
        "title": "InvoiceSummary",
        "type": "object"
      },
...
      "PayInvoiceRequest": {
        "additionalProperties": false,
        "description": "Choose wallet funding; the backend determines all payment amounts.",
        "properties": {
          "use_wallet": {
            "default": true,
            "title": "Use Wallet",
            "type": "boolean"
          }
        },
        "title": "PayInvoiceRequest",
        "type": "object"
      },
      "PayInvoiceResponse": {
        "description": "The result of paying an invoice.\n\n``status`` is ``PAID`` when the wallet covered the total or development settles the\ngateway portion directly. ``PENDING`` requires a gateway payment and has ``payment_url``.",
        "properties": {
          "amount_from_gateway": {
            "title": "Amount From Gateway",
            "type": "string"
          },
          "amount_from_wallet": {
            "title": "Amount From Wallet",
            "type": "string"
          },
          "invoice_id": {
            "title": "Invoice Id",
            "type": "string"
          },
          "payment_intent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Intent Id"
          },
          "payment_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Url"
          },
          "session_reused": {
            "default": false,
            "title": "Session Reused",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "use_wallet": {
            "default": true,
            "title": "Use Wallet",
            "type": "boolean"
          }
        },
        "required": [
          "invoice_id",
          "status",
          "amount_from_wallet",
          "amount_from_gateway"
        ],
        "title": "PayInvoiceResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Giftly API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/invoices": {
      "get": {
        "description": "List owned invoice revisions; issued-date bounds use Asia/Riyadh days.",
        "operationId": "list_invoices_api_invoices_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvoiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "include_historical",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Historical",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Inclusive start date (Gregorian YYYY-MM-DD).",
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start date (Gregorian YYYY-MM-DD).",
              "title": "From Date"
            }
          },
          {
            "description": "Inclusive end date (Gregorian YYYY-MM-DD).",
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive end date (Gregorian YYYY-MM-DD).",
              "title": "To Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Invoices",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/invoices/{invoice_id}": {
      "get": {
        "description": "Return an invoice by id (participant only).",
        "operationId": "get_invoice_api_invoices__invoice_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/invoices/{invoice_id}/cancel": {
      "post": {
        "description": "Cancel an unpaid issued invoice, reopening the order (issuing courier only).",
        "operationId": "cancel_invoice_api_invoices__invoice_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/invoices/{invoice_id}/pay": {
      "post": {
        "description": "Pay an issued invoice from wallet, gateway, or a split of both (customer only).",
        "operationId": "pay_invoice_api_invoices__invoice_id__pay_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PayInvoiceRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayInvoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pay Invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/invoices/{invoice_id}/pdf": {
      "get": {
        "description": "Download a stored invoice as English PDF, scoped to order participants.",
        "operationId": "download_invoice_pdf_api_invoices__invoice_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Invoice Pdf",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/invoices/{invoice_id}/promo": {
      "post": {
        "description": "Apply/remove a promo by replacing an unpaid invoice; pay the returned invoice ID.",
        "operationId": "apply_invoice_promo_api_invoices__invoice_id__promo_post",
        "parameters": [
          {
            "in": "path",
            "name": "invoice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invoice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyInvoicePromoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Invoice Promo",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/api/orders/{order_id}/invoice": {
      "get": {
        "description": "Return an order's active invoice (participant only).",
        "operationId": "get_order_invoice_api_orders__order_id__invoice_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order Invoice",
        "tags": [
          "invoices"
        ]
      }
    },
    "/api/orders/{order_id}/invoices": {
      "post": {
        "description": "Author and issue an invoice for an order (assigned courier only).",
        "operationId": "create_invoice_api_orders__order_id__invoices_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invoice",
        "tags": [
          "invoices"
        ]
      }
    },
...
  }
}